kfac/requirements_cpu.txt
jax[cpu]
tensorflow-cpu
tensorflow-datasets
tensorboardX
matplotlib
argparse
parse
pandas
seaborn
